Introduce ZSharedPtr and ZUniquePtr into the whole system  to manage object lifetime. Add gender and 2nd password setting for WvsLogin. Add duplicated id check when creating characters. Four basic attributes : STR/DEX/INT/LUK are now read from packet sent from the client. Use "LoadAll" to load items instead of loading them separately. Solved several memory leak, the most major part is memory leak from scripts and drops. Add Party2 script funcs, and had it tested. Add change Hair/Skin/Face scripts. Correct the type of time integers(which should be unsigned). Add some funcs in ScriptUser/ScriptInventory. Fix some bugs in Trading. Fix some bugs in PersonalShop. Portals/Quests are now able to support integrated scripts(see portal/market.lua).

+++ WvsGame/AffectedArea.cpp
@@ -20,7 +20,7 @@ void AffectedArea::MakeEnterFieldPacket(OutPacket * oPacket)
 	oPacket->Encode4(m_nOwnerID);
 	oPacket->Encode4(m_nSkillID);
 	oPacket->Encode1((char)m_nSLV);
-	int tTime = (m_tStart - GameDateTime::GetTime()) / 100;
+	int tTime = (int)(((long long int)m_tStart - (long long int)GameDateTime::GetTime()) / 100);
 	if (tTime < 0)
 		tTime = 0;
 	oPacket->Encode2((short)tTime);
